+++ src/frontend/web_application/src/scenes/NewDraft/presenter.jsx
@@ -22,6 +22,10 @@ class NewDraft extends Component {
     internalId: undefined,
   };
 
+  state = {
+    isSending: false,
+  };
+
   componentDidMount() {
     const { internalId, draft, requestNewDraft } = this.props;
     if (!internalId || !draft) {
@@ -34,6 +38,15 @@ class NewDraft extends Component {
 
     return action({ internalId, draft, message });
   };
+
+  handleSend = ({ draft }) => {
+    const { sendDraft, internalId, message } = this.props;
+    const params = { draft, message, internalId };
+
+    this.setState({ isSending: true });
+
+    return sendDraft(params).then(() => this.setState({ isSending: false }));
+  }
 
   handleDelete = () => {
     const { message, onDeleteMessage } = this.props;
@@ -53,7 +66,7 @@ class NewDraft extends Component {
   }
 
   render() {
-    const { draft, internalId, editDraft, onSaveDraft, sendDraft } = this.props;
+    const { draft, internalId, editDraft, onSaveDraft } = this.props;
 
     return (
       <ScrollToWhenHash forceTop>
@@ -62,7 +75,8 @@ class NewDraft extends Component {
           draft={draft}
           onChange={this.makeHandle(editDraft)}
           onSave={this.makeHandle(onSaveDraft)}
-          onSend={this.makeHandle(sendDraft)}
+          onSend={this.handleSend}
+          isSending={this.state.isSending}
           renderDraftMessageActionsContainer={this.renderDraftMessageActionsContainer}
         />
       </ScrollToWhenHash>